move code to ips.go
